Move all pg_attribute logic to PgAttributeImpl.
